requirements.txt
certifi==2024.6.2
pymongo==4.7.2
python-dotenv==1.0.1